ComfyUI-master-fitow/custom_nodes/Google-Gemini-Concurrent/requirements.txt
google-genai>=1.51.0
Pillow
torch